improve data. extract BiomeGeneratorHelper
