--- src/main/scala/hohserg/dimensional/layers/data/layer/base/DimensionalGenerator.scala
@@ -1,9 +1,0 @@
-package hohserg.dimensional.layers.data.layer.base
-
-import hohserg.dimensional.layers.worldgen.proxy.server.ProxyWorldServer
-
-trait DimensionalGenerator extends Generator {
-
-  def proxyWorld: ProxyWorldServer
-
-}
