+++ src/main/scala/hohserg/dimensional/layers/data/layer/base/Layer.scala
@@ -1,7 +1,10 @@
 package hohserg.dimensional.layers.data.layer.base
 
 import hohserg.dimensional.layers.preset.LayerSpec
-import hohserg.dimensional.layers.{CCWorld, CCWorldServer}
+import hohserg.dimensional.layers.worldgen.proxy.client.ProxyWorldClient
+import hohserg.dimensional.layers.{CCWorld, CCWorldClient, CCWorldServer}
+import net.minecraft.world.DimensionType
+import net.minecraftforge.fml.relauncher.{Side, SideOnly}
 
 trait Layer {
   type Spec <: LayerSpec
@@ -21,4 +24,21 @@ trait Layer {
     case serverWorld: CCWorldServer => createGenerator(serverWorld)
   }
 
+}
+
+trait DimensionalLayer extends Layer {
+  override type Bounds = DimensionalLayerBounds
+  override type G <: DimensionalGenerator
+
+  def dimensionType: DimensionType
+
+  def isCubic: Boolean
+
+  @SideOnly(Side.CLIENT)
+  protected def createClientProxyWorld(original: CCWorldClient): ProxyWorldClient = ProxyWorldClient(original, this)
+
+  @SideOnly(Side.CLIENT)
+  lazy val clientProxyWorld: ProxyWorldClient = originalWorld match {
+    case clientWorld: CCWorldClient => createClientProxyWorld(clientWorld)
+  }
 }
